Refactor : Redis 코드 리팩터링
