+++ src/configs/index.ts
@@ -45,7 +45,6 @@ const fcmConfig = {
 const redisConfig = {
   host: process.env.REDIS_HOST,
   port: Number(process.env.REDIS_PORT),
-  password: process.env.REDIS_PASSWORD,
 } as RedisConfig;
 
 // ENV Module
